requirements.txt
google-generativeai==0.4.1
pillow==10.2.0